updating cloud with joinrole

+++ functions/guildmemberadd.js
@@ -18,11 +18,21 @@ module.exports = {
 
         if (permission.check_if_valid_author(msg, db)) {
             try {
-                if (db.prepare(`SELECT id FROM onJoin WHERE id='${msg.guild.id}'`).get()) {
-                    db.prepare(`DELETE FROM onJoin WHERE id='${msg.guild.id}'`).run()
+                if(!args[0]) {
+                    msg.react("❌");
+                    return;
                 }
-                db.prepare(`INSERT INTO onJoin VALUES(${msg.guild.id},'${args[0]}')`).run();
-                msg.react("✅");
+                if(msg.guild.roles.fetch(args[0])){
+                    if (db.prepare(`SELECT id FROM onJoin WHERE id='${msg.guild.id}'`).get()) {
+                        db.prepare(`DELETE FROM onJoin WHERE id='${msg.guild.id}'`).run()
+                    }
+                    db.prepare(`INSERT INTO onJoin VALUES(${msg.guild.id},'${args[0]}')`).run();
+                    msg.react("✅");
+                }
+                else {
+                    msg.react("❌");
+                }
+                
             }
             catch (err) {
                 console.log(err);
@@ -54,16 +64,21 @@ module.exports = {
             msg.react("❌");
         }
     },
-    list: async function (msg, db) {
+    list: async function (msg, db, client) {
         if (db.prepare(`SELECT id FROM onJoin WHERE id='${msg.guild.id}'`).get()) {
             let response = db.prepare(`SELECT * FROM 'onJoin' WHERE id='${msg.guild.id}'`).get();
             let roleID = response['roleID'];
             let role = await msg.guild.roles.fetch(roleID);
-            msg.channel.send(
-                new Discord.MessageEmbed().setTitle("Role Given when Joined").setDescription(role.toString()))
+            if(role){
+                msg.channel.send(new Discord.MessageEmbed().setTitle("Role Given when Joined").setDescription(role.toString()))
+            }
+            else {
+                msg.channel.send("There was an Error");
+            }
+            
         }
         else {
-            msg.channel.send(new Discord.MessageEmbed().setTitle("No Join Role Configured").setDescription(`Check out \`${config.prefix} help configure\` for help`))
+            msg.channel.send(new Discord.MessageEmbed().setTitle("No Join Role Configured").setDescription(`Check out \`${config.prefix}help configure\` for help`).setThumbnail(client.user.avatarURL()))
         }
     }
 }
